[2017.12.08 정영현] profile image 에러 나는 부분 수정하고 서버쪽 이미지 파일이 삭제되는 경우가 있어서 에러 발생 관련된 class 추가 합니다.

+++ SIR/src/net/profileimage/controller/DeleteImageFile.java
@@ -22,11 +22,13 @@ public class DeleteImageFile {
 			String fileName = pathArray[pathArray.length-1];
 			
 			ServletContext ctx = request.getServletContext();
-			/*업로드할 실제 서버경로 얻기 - upload 폴더명 맞추기*/
-			String realPath = ctx.getRealPath("uploadProfileImage/");
+			String folderPath = ctx.getRealPath("uploadProfileImage/");
+			String transForderPath = folderPath.replace("\\", "\\\\");
 	
-			File delFile = new File(realPath + fileName);
-			delFile.delete();	
+			File delFile = new File(transForderPath + fileName);
+			if (delFile.exists()){
+				delFile.delete();	
+			}
 		}
 	}
 }
